# public/swagger/swagger.yaml
openapi: '3.0.0'
info:
  title: Topmama Book API's
  description: Topmama books API
  version: 1.0.0
  contact: 
    name: Donald O
    email: donopiyo@gmail.com
servers:
- url: http://local-dev:8000/
  description: kvm development server
...
paths:
...
  /api/v1/books/{id}/comments:
    summary: Book comments operations
    description: Returns a book's comments based on id
    get:
      summary: Retrieve a book's comments
      operationId: bookComments
      tags:
        - Book
        - Comments
      parameters:
        - name: id
          in: path
          description: book unique identity
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: A list of book's comments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentWithId'
                  
    post:
      tags:
        - Comments
        - Book
      summary: Adding a new comment
      operationId: createBookComment
      parameters:
        - name: id
          in: path
          description: book unique identity
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
      responses:
        '201':
          description: Created              
  /api/v1/books/{id}/characters:
    summary: Book characters operation
    description: Returns a book's characters based on id
    get:
      summary: retrieve a book's characters
      tags:
        - Book
        - Characters
      operationId: bookCharacters
      parameters:
        - name: id
          in: path
          description: book unique identity
          required: true
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/sortCharacters'
        - $ref: '#/components/parameters/sortTypeCharacters'
        - $ref: '#/components/parameters/filterCharacters'
          
      responses:
        '200':
          description: A list of a book's characters.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Character'
    post:
      tags:
        - Book
        - Characters
      summary: Adding a new character
      operationId: createBookCharacter
      parameters:
        - name: id
          in: path
          description: character unique identity
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Character'
      responses:
        '201':
          description: Created
...
components:
  schemas:
...
    Character:
      type: object
      required: 
        - name
        - gender
      properties:
        name:
          type: string
        age:
          type: integer
          format: int32
        gender:
          $ref: '#/components/schemas/Gender'
...
    Comment:
      type: object
      required:
        - comment
      properties:
        comment:
          type: string
          maxLength: 500
    CommentWithId:
      allOf:
        - $ref: '#/components/schemas/Comment'
        - type: object
          properties:
            id:
              type: integer
              format: int64
            ip:
              type: string
            created_at:
              type: string
            updated_at:
              type: string
    Gender:
      type: object
      required:
        - gender_type
      properties:
        gender_type:
          type: string
  parameters:
    sortCharacters:
      name: sortby
      in: query
      description: sort paraemter
      required: false
      schema:
        type: string
        enum: [name,age,gender]
    sortTypeCharacters:
      name: direction
      in: query
      description: in ascending or descending order
      required: false
      schema:
        type: string
        enum: [asc,desc]
    filterCharacters:
      name: filter
      in: query
      description: filter paraemter
      required: false
      schema:
        type: string
        enum: [Male,Female,Other]
...